Moved packages that relate to PlayerBoard from package stone_age to player_board
