+++ src/main/java/sk/uniba/fmph/dcs/player_board/PlayerBoardFacade.java
@@ -1,11 +1,13 @@
-package sk.uniba.fmph.dcs.stone_age;
+package sk.uniba.fmph.dcs.player_board;
+
+import sk.uniba.fmph.dcs.stone_age.*;
 
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.List;
 import java.util.OptionalInt;
 
-public class PlayerBoardFacade implements InterfacePlayerBoardGameBoard,InterfaceNewTurn,InterfaceFeedTribe {
+public class PlayerBoardFacade implements InterfacePlayerBoardGameBoard, InterfaceNewTurn, InterfaceFeedTribe {
     private PlayerBoard playerBoard;
     public PlayerBoardFacade(){
         this.playerBoard = new PlayerBoard();
