+++ src/main/java/sk/uniba/fmph/dcs/player_board/TribeFedStatus.java
@@ -1,4 +1,6 @@
-package sk.uniba.fmph.dcs.stone_age;
+package sk.uniba.fmph.dcs.player_board;
+
+import sk.uniba.fmph.dcs.stone_age.Effect;
 
 import java.util.ArrayList;
 import java.util.Collection;
